refactor: set basic header
