+++ src/assets/style/App.scss
@@ -1,3 +1,5 @@
+@import './shared.scss';
+
 * {
   margin: 0;
   padding: 0;
@@ -5,23 +7,22 @@
   overscroll-behavior: none;
 }
 
-body {
+html,
+body,
+#__next {
+  width: 100%;
+  height: 100%;
   position: relative;
 }
 
-$color-main: #ff4d40;
-$color-black: #333;
-$color-white: #fff;
-$color-lightgray: #f7f7f7;
-$color-gray: #efefef;
-$color-darkgray: #c3c3c3;
+.app-outer-container {
+  width: 100%;
+  min-height: calc(100% - 64px); // header height
 
-$padding-base: 14px;
-$padding-width: calc(100% - 28px);
+  margin-top: $header-height;
 
-$margin-base: 14px;
-$margin-width: calc(100% - 28px);
-
-$box-border-radius: 8px;
-
-$icon-font-size: 24px;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+}
