+++ src/assets/style/Main.scss
@@ -2,7 +2,6 @@
 
 .main-feed-container {
   width: 100%;
-  min-height: 220px;
   height: 30vh;
   max-height: 280px;
 
@@ -90,6 +89,12 @@
           border-bottom: none;
           transition: all 0.3s;
         }
+      }
+    }
+
+    .feed-tab-date {
+      @media screen and (max-width: 365px) {
+        display: none;
       }
     }
   }
