+++ src/assets/style/shared.scss
@@ -1,0 +1,20 @@
+// app theme style
+$color-main: #ff4d40;
+$color-black: #333;
+$color-white: #fff;
+$color-lightgray: #f7f7f7;
+$color-gray: #efefef;
+$color-darkgray: #c3c3c3;
+
+$padding-base: 14px;
+$padding-width: calc(100% - 28px);
+
+$margin-base: 14px;
+$margin-width: calc(100% - 28px);
+
+$box-border-radius: 8px;
+
+$icon-font-size: 24px;
+// ----------------------------------
+
+$header-height: 64px;
